update temporary creds for S3 in `AWS Distributor` only
